+++ inlang/source-code/ide-extension/src/commands/openInEditor.ts
@@ -6,7 +6,7 @@ import { getGitOrigin } from "../utilities/settings/getGitOrigin.js"
 
 export const openInEditorCommand = {
 	command: "sherlock.openInEditor",
-	title: "Inlang: Open in Editor",
+	title: "Sherlock: Open in Editor",
 	register: commands.registerCommand,
 	callback: async function (args: { messageId: Message["id"]; selectedProjectPath: string }) {
 		// TODO: Probably the origin should be configurable via the config.