+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/adapter-sveltekit/examples/lang-in-slug-static/src/routes/[lang]/about/+layout.server.ts
@@ -1,6 +1,0 @@
-import { initServerLoadWrapper } from "@inlang/paraglide-js-sveltekit/adapter-sveltekit/server"
-import type { LayoutServerLoad } from "./$types.js"
-
-export const load = initServerLoadWrapper<LayoutServerLoad>().use(async (_, { i }) => {
-	console.info("[lang]/about/+layout.server.ts", i("welcome"))
-})