+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/adapter-sveltekit/examples/test-app/src/routes/[lang]/+layout.server.js
@@ -1,3 +1,0 @@
-export const load = () => {
-	return { "layout.server.js": 1 }
-}