+++ inlang/source-code/paraglide/paraglide-js-sveltekit-old/src/detectors/detectors/client/navigatorDetector.ts
@@ -1,7 +1,0 @@
-import type { Detector } from "../../types.js"
-
-/**
- * Detects an array of languageTags based on the user preferences stored in window.navigator.languageTags
- * @returns An Array of strings representing the users preferred languageTags ordered after preference.
- */
-export const navigatorDetector = (() => [...window.navigator.languages]) satisfies Detector