+++ inlang/source-code/website/src/pages/m/@uid/@id/+Page.tsx
@@ -477,12 +477,7 @@ const EcosystemIncompatibleBadgeBig = () => {
 export function Recommends(props: { recommends: MarketplaceManifest[] }) {
 	return (
 		<>
-			<Show
-				when={currentPageContext.urlParsed.pathname.includes("/m/")}
-				fallback={<h3 class="font-semibold mb-4">Read more:</h3>}
-			>
-				<h3 class="font-semibold mb-4">Recommended to use with:</h3>
-			</Show>
+			<h3 class="font-semibold mb-4">Recommended:</h3>
 			<div class="flex items-center gap-4 md:flex-row flex-col">
 				<For each={props.recommends}>
 					{/* @ts-ignore */}
@@ -513,11 +508,14 @@ function NavbarCommon(props: {
 			.replaceAll("(", "")
 			.replaceAll(")", "")
 			.replaceAll("?", "")
+			.replaceAll("!", "")
 			.replaceAll(".", "")
 			.replaceAll("@", "")
 			.replaceAll(/([\uE000-\uF8FF]|\uD83C[\uDF00-\uDFFF]|\uD83D[\uDC00-\uDDFF])/g, "")
 			.replaceAll("✂", "")
 			.replaceAll(":", "")
+			.replaceAll("'", "")
+			.replaceAll("&", "")
 	}
 
 	const isSelected = (heading: string) => {
@@ -542,9 +540,10 @@ function NavbarCommon(props: {
 							: new Promise((resolve) => img.addEventListener("load", resolve))
 					)
 				)
-
-				//scrollToAnchor(replaceChars(sectionTitle.toString().toLowerCase()), "smooth")
-				setHighlightedAnchor(replaceChars(sectionTitle.toString().toLowerCase()))
+				setTimeout(() => {
+					scrollToAnchor(replaceChars(sectionTitle.toString().toLowerCase()), "smooth")
+					setHighlightedAnchor(replaceChars(sectionTitle.toString().toLowerCase()))
+				}, 100)
 			} else {
 				for (const heading of props.tableOfContents[sectionTitle]!) {
 					if (
@@ -559,9 +558,10 @@ function NavbarCommon(props: {
 									: new Promise((resolve) => img.addEventListener("load", resolve))
 							)
 						)
-
-						scrollToAnchor(replaceChars(heading.toString().toLowerCase()), "smooth")
-						setHighlightedAnchor(replaceChars(heading.toString().toLowerCase()))
+						setTimeout(() => {
+							scrollToAnchor(replaceChars(heading.toString().toLowerCase()), "smooth")
+							setHighlightedAnchor(replaceChars(heading.toString().toLowerCase()))
+						}, 100)
 					}
 				}
 			}